[#264] multipart type을 http 하위 모듈로 이동
